+++ packages/synapse-interface/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.35](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.1.34...@synapsecns/synapse-interface@0.1.35) (2023-06-28)
+
+**Note:** Version bump only for package @synapsecns/synapse-interface
+
+
+
+
+
+## [0.1.34](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.1.33...@synapsecns/synapse-interface@0.1.34) (2023-06-27)
+
+**Note:** Version bump only for package @synapsecns/synapse-interface
+
+
+
+
+
 ## [0.1.33](https://github.com/synapsecns/sanguine/compare/@synapsecns/synapse-interface@0.1.32...@synapsecns/synapse-interface@0.1.33) (2023-06-20)
 
 **Note:** Version bump only for package @synapsecns/synapse-interface